Rename component files to follow PascalCase naming convention (#79)
* Update CallToAction.yml to correct datasource template path for CTA Sections

* Refactor component names in component maps to lowercase for consistency in location finder and product listing examples

* Update CustomerTestimonials.yml to correct datasource template path for Testimonials

* Enhance Testimonials component to handle missing datasource gracefully with NoDataFallback and improve optional chaining for safety

* Refactor component names in YAML files and component maps to use PascalCase for consistency across the project

* Refactor Testimonials component to simplify datasource validation by removing unnecessary checks for title and tagline

* Refactor component imports in component map to use PascalCase for consistency across the project

* Rename testimonials.tsx to Testimonials.tsx for consistent PascalCase naming

* Rename calltoaction.tsx to CallToAction.tsx and hero.tsx to Hero.tsx for consistent PascalCase naming

* Rename calltoaction.tsx to CallToAction.tsx, hero.tsx to Hero.tsx, and testimonials.tsx to Testimonials.tsx for consistent PascalCase naming

* Update component imports in component map to use PascalCase for CallToAction, Hero, and Testimonials for consistency
