+++ authoring/items/items/templates/items/ccl.renderings/click-click-launch/Promos/CustomerTestimonials.yml
@@ -6,7 +6,7 @@ Path: "/sitecore/layout/Renderings/Project/click-click-launch/Promos/CustomerTes
 SharedFields:
 - ID: "037fe404-dd19-4bf7-8e30-4dadf68b27b0"
   Hint: componentName
-  Value: testimonials
+  Value: Testimonials
 - ID: "06d5295c-ed2f-4a54-9bf2-26228d113318"
   Hint: __Icon
   Value: Office/32x32/messages.png
@@ -50,7 +50,7 @@ SharedFields:
     
 - ID: "1a7c85e5-dc0b-490d-9187-bb1dbcb4c72f"
   Hint: Datasource Template
-  Value: "/sitecore/templates/Project/xmc-cl/Component Library/Testimonials/Customer Testimonial Datasource"
+  Value: "/sitecore/templates/Project/click-click-launch/Components/Testimonials/Customer Testimonial Datasource"
 - ID: "7d8ae35f-9ed1-43b5-96a2-0a5f040d4e4e"
   Hint: Open Properties after Add
   Value: 0
