+++ examples/kit-nextjs-location-finder/.sitecore/component-map.ts
@@ -10,11 +10,11 @@ import * as ArticleHeader from 'components/article-header/ArticleHeader';
 import * as BackgroundThumbnail from 'components/background-thumbnail/BackgroundThumbnail.dev';
 import * as Breadcrumbs from 'components/breadcrumbs/Breadcrumbs';
 import * as ButtonComponent from 'components/button-component/ButtonComponent';
-import * as CallToAction from 'components/component-library/calltoaction';
+import * as CallToAction from 'components/component-library/CallToAction';
 import * as Card from 'components/card/Card.dev';
 import * as CardSpotlight from 'components/card-spotlight/card-spotlight.dev';
 import * as Carousel from 'components/carousel/Carousel';
-import * as CLHero from 'components/component-library/hero';
+import * as CLHero from 'components/component-library/Hero';
 import * as ColumnSplitter from 'components/sxa/ColumnSplitter';
 import * as ContactSection from 'components/component-library/ContactSection';
 import * as Container from 'components/sxa/Container';
@@ -80,7 +80,7 @@ import * as SubmissionForm from 'components/submission-form/SubmissionForm';
 import * as SubscriptionBanner from 'components/subscription-banner/SubscriptionBanner';
 import * as TeamSection from 'components/component-library/TeamSection';
 import * as TestimonialCarousel from 'components/testimonial-carousel/TestimonialCarousel';
-import * as Testimonials from 'components/component-library/testimonials';
+import * as Testimonials from 'components/component-library/Testimonials';
 import * as TextBanner from 'components/text-banner/TextBanner';
 import * as ThemeProvider from 'components/theme-provider/theme-provider.dev';
 import * as Title from 'components/sxa/Title';
