+++ examples/kit-nextjs-location-finder/src/components/component-library/Hero.tsx
