+++ examples/kit-nextjs-product-listing/src/components/component-library/Hero.tsx
