+++ examples/kit-nextjs-product-listing/src/components/component-library/Testimonials.tsx
@@ -20,6 +20,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faStar } from '@fortawesome/free-solid-svg-icons';
 import { faStar as faStarOutline } from '@fortawesome/free-regular-svg-icons';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from 'shadcd/components/ui/tabs';
+import { NoDataFallback } from '@/utils/NoDataFallback';
 
 interface Fields {
   data: {
@@ -254,7 +255,11 @@ const TestimonialCard = (props: TestimonialCardProps) => {
 };
 
 export const Default = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -268,7 +273,7 @@ export const Default = (props: TestimonialsProps): JSX.Element => {
           </p>
         </div>
         <div className="flex flex-col md:flex-row gap-x-12 gap-y-20">
-          {datasource.children.results.map((testimonial) => (
+          {datasource.children?.results?.map((testimonial) => (
             <TestimonialCard
               key={testimonial.id}
               testimonial={testimonial}
@@ -276,7 +281,7 @@ export const Default = (props: TestimonialsProps): JSX.Element => {
               withLogo
               className="flex-1"
             />
-          ))}
+          )) || []}
         </div>
       </div>
     </section>
@@ -284,7 +289,11 @@ export const Default = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials1 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -300,11 +309,11 @@ export const Testimonials1 = (props: TestimonialsProps): JSX.Element => {
         <div className="max-w-5xl mx-auto px-12">
           <Carousel opts={{ loop: true }} className="w-full">
             <CarouselContent>
-              {datasource.children.results.map((testimonial) => (
+              {datasource.children?.results?.map((testimonial) => (
                 <CarouselItem key={testimonial.id}>
                   <TestimonialCard testimonial={testimonial} type="centered" withLogo withRating />
                 </CarouselItem>
-              ))}
+              )) || []}
             </CarouselContent>
             <CarouselPrevious className="disabled:hidden" />
             <CarouselNext className="disabled:hidden" />
@@ -316,7 +325,11 @@ export const Testimonials1 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials2 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -332,11 +345,11 @@ export const Testimonials2 = (props: TestimonialsProps): JSX.Element => {
 
         <Carousel opts={{ align: 'start', loop: true }} className="w-full">
           <CarouselContent>
-            {datasource.children.results.map((testimonial) => (
+            {datasource.children?.results?.map((testimonial) => (
               <CarouselItem key={testimonial.id} className="pr-4 md:basis-1/2">
                 <TestimonialCard testimonial={testimonial} type="simple" withLogo withRating />
               </CarouselItem>
-            ))}
+            )) || []}
           </CarouselContent>
           <div className="flex items-center gap-2 mt-8">
             <CarouselPrevious className="static translate-0 disabled:hidden" />
@@ -349,7 +362,11 @@ export const Testimonials2 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials3 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -366,7 +383,7 @@ export const Testimonials3 = (props: TestimonialsProps): JSX.Element => {
 
           <Carousel opts={{ align: 'start', loop: true }} className="w-full">
             <CarouselContent>
-              {datasource.children.results.map((testimonial) => (
+              {datasource.children?.results?.map((testimonial) => (
                 <CarouselItem key={testimonial.id} className="pr-2 md:basis-3/4">
                   <TestimonialCard
                     testimonial={testimonial}
@@ -375,7 +392,7 @@ export const Testimonials3 = (props: TestimonialsProps): JSX.Element => {
                     className="h-full"
                   />
                 </CarouselItem>
-              ))}
+              )) || []}
             </CarouselContent>
             <div className="flex items-center gap-2 mt-8">
               <CarouselPrevious className="static translate-0" />
@@ -389,7 +406,11 @@ export const Testimonials3 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials4 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -406,11 +427,11 @@ export const Testimonials4 = (props: TestimonialsProps): JSX.Element => {
 
           <Carousel opts={{ align: 'start', loop: true }} className="w-full">
             <CarouselContent>
-              {datasource.children.results.map((testimonial) => (
+              {datasource.children?.results?.map((testimonial) => (
                 <CarouselItem key={testimonial.id}>
                   <TestimonialCard testimonial={testimonial} type="boxed" withRating />
                 </CarouselItem>
-              ))}
+              )) || []}
             </CarouselContent>
             <div className="flex items-center justify-end gap-2 mt-8">
               <CarouselPrevious className="static translate-0 disabled:hidden" />
@@ -424,7 +445,11 @@ export const Testimonials4 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials5 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -440,11 +465,11 @@ export const Testimonials5 = (props: TestimonialsProps): JSX.Element => {
 
         <Carousel opts={{ align: 'start', loop: true }} className="w-full">
           <CarouselContent>
-            {datasource.children.results.map((testimonial) => (
+            {datasource.children?.results?.map((testimonial) => (
               <CarouselItem key={testimonial.id}>
                 <TestimonialCard testimonial={testimonial} type="large" withLogo withRating />
               </CarouselItem>
-            ))}
+            )) || []}
           </CarouselContent>
           <div className="flex items-center justify-end gap-2 mt-8">
             <CarouselPrevious className="static translate-0 disabled:hidden" />
@@ -457,7 +482,11 @@ export const Testimonials5 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials6 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -471,7 +500,7 @@ export const Testimonials6 = (props: TestimonialsProps): JSX.Element => {
           </p>
         </div>
         <div className="md:columns-3 gap-8">
-          {datasource.children.results.map((testimonial) => (
+          {datasource.children?.results?.map((testimonial) => (
             <TestimonialCard
               key={testimonial.id}
               testimonial={testimonial}
@@ -479,7 +508,7 @@ export const Testimonials6 = (props: TestimonialsProps): JSX.Element => {
               withRating
               className="mb-8"
             />
-          ))}
+          )) || []}
         </div>
       </div>
     </section>
@@ -487,7 +516,11 @@ export const Testimonials6 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials7 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -501,7 +534,7 @@ export const Testimonials7 = (props: TestimonialsProps): JSX.Element => {
           </p>
         </div>
         <div className="flex flex-col md:flex-row gap-x-12 gap-y-20">
-          {datasource.children.results.map((testimonial) => (
+          {datasource.children?.results?.map((testimonial) => (
             <TestimonialCard
               key={testimonial.id}
               testimonial={testimonial}
@@ -509,7 +542,7 @@ export const Testimonials7 = (props: TestimonialsProps): JSX.Element => {
               withRating
               className="flex-1"
             />
-          ))}
+          )) || []}
         </div>
       </div>
     </section>
@@ -517,7 +550,11 @@ export const Testimonials7 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials8 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -533,7 +570,7 @@ export const Testimonials8 = (props: TestimonialsProps): JSX.Element => {
 
         <Carousel opts={{ align: 'start', loop: true }} className="w-full">
           <CarouselContent>
-            {datasource.children.results.map((testimonial) => (
+            {datasource.children?.results?.map((testimonial) => (
               <CarouselItem key={testimonial.id} className="md:basis-1/2">
                 <TestimonialCard
                   testimonial={testimonial}
@@ -543,7 +580,7 @@ export const Testimonials8 = (props: TestimonialsProps): JSX.Element => {
                   className="h-full"
                 />
               </CarouselItem>
-            ))}
+            )) || []}
           </CarouselContent>
           <div className="flex items-center justify-end gap-2 mt-8">
             <CarouselPrevious className="static translate-0 disabled:hidden" />
@@ -556,7 +593,11 @@ export const Testimonials8 = (props: TestimonialsProps): JSX.Element => {
 };
 
 export const Testimonials9 = (props: TestimonialsProps): JSX.Element => {
-  const datasource = useMemo(() => props.fields.data.datasource, [props.fields.data.datasource]);
+  const datasource = useMemo(() => props.fields.data?.datasource, [props.fields.data?.datasource]);
+
+  if (!datasource) {
+    return <NoDataFallback componentName="Testimonials" />;
+  }
 
   return (
     <section className={`py-24 px-4 ${props.params.styles}`} data-class-change>
@@ -572,7 +613,7 @@ export const Testimonials9 = (props: TestimonialsProps): JSX.Element => {
 
         <Tabs defaultValue={datasource.children.results[0].id} className="mt-20">
           <TabsList>
-            {datasource.children.results.map((testimonial) => (
+            {datasource.children?.results?.map((testimonial) => (
               <TabsTrigger value={testimonial.id} key={testimonial.id}>
                 <ContentSdkImage
                   field={testimonial.testimonialIcon?.jsonValue}
@@ -581,14 +622,14 @@ export const Testimonials9 = (props: TestimonialsProps): JSX.Element => {
                   className="object-contain"
                 />
               </TabsTrigger>
-            ))}
+            )) || []}
           </TabsList>
 
-          {datasource.children.results.map((testimonial) => (
+          {datasource.children?.results?.map((testimonial) => (
             <TabsContent value={testimonial.id} key={testimonial.id} className="py-16">
               <TestimonialCard testimonial={testimonial} type="centered" withRating />
             </TabsContent>
-          ))}
+          )) || []}
         </Tabs>
       </div>
     </section>
